refactord CLI as a module
