+++ prp/cli/__init__.py
@@ -1,0 +1,3 @@
+"""Definition of CLI commands."""
+
+from .base import cli
